+++ src/app/components/products/product/product.component.ts
@@ -1,0 +1,14 @@
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
+
+import { ProductModel } from '../../../models/product.model';
+
+@Component({
+  selector: 'app-product',
+  templateUrl: './product.component.html',
+  styleUrls: ['./product.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class ProductComponent {
+  @Input() product: ProductModel;
+  @Output() buy = new EventEmitter<number>();
+}