+++ src/app/models/product.model.ts
@@ -13,5 +13,5 @@ export interface ProductModel {
   readonly price: number;
   readonly category: Categorys;
   readonly isAvailable: boolean;
-  quantity?: number;
+  readonly quantity?: number;
 }