+++ src/app/shared/shared.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SharedComponent } from './shared.component';
+import { HighlightDirective } from './directives/highlight.directive';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [SharedComponent, HighlightDirective],
+  exports: [HighlightDirective]
+})
+export class SharedModule { }